+++ assets/js/wpdk-controls.js
@@ -68,6 +68,43 @@ if ( typeof( jQuery.fn.swipe ) === 'undefined' ) {
 }
 
 /**
+ * WPDK Selectable Checkbox by shift key
+ *
+ * Usage: $form.find('input[type="checkbox"]').wpdkShiftSelectableCheckbox();
+ * replace input[type="checkbox"] with the selector to match your list of checkboxes
+ *
+ */
+if ( typeof jQuery.fn.wpdkShiftSelectableCheckbox === 'undefined' ) {
+  +function ( $ )
+  {
+
+    'use strict';
+
+    $.fn.wpdkShiftSelectableCheckbox = function ()
+    {
+      var lastChecked, $boxes = this;
+
+      $boxes.click( function ( evt )
+      {
+        if ( !lastChecked ) {
+          lastChecked = this;
+          return;
+        }
+
+        if ( evt.shiftKey ) {
+          var start = $boxes.index( this ), end = $boxes.index( lastChecked );
+          $boxes.slice( Math.min( start, end ), Math.max( start, end ) + 1 )
+            .attr( 'checked', lastChecked.checked )
+            .trigger( 'change' );
+        }
+
+        lastChecked = this;
+      } );
+    };
+  }( jQuery );
+}
+
+/**
  * This class init all WPDK controls, attach new event and perform special actions.
  *
  * @class           WPDKControls